app/pyproject.toml
[tool.poetry]
name = "TEMPLATEPROJECT"
version = "0.1.0"
description = "TEMPLATEPROJECT"
authors = ["systems@hactar.is"]


[tool.poetry.dependencies]
python = ">=3.6,<3.9"

wagtail = "2.9"
arrow = "^0.15.5"
bugsnag = "^3.6"
envkey = "^1.2"
loguru = "^0.4.1"
toml = "^0.10.0"
hiredis = "^1.0"
django-redis = "^4.10"
django-jinja = "^2.4"
django-cacheops = "^4.1"
django-extensions = "^2.2"
wagtail-modelchooser = "^2.2.2"
wagtailfontawesome = "^1.1"
django-autocomplete-light = "^3.4"
django-storages = "^1.7"
psycopg2 = "^2.8"
ipython = "^7.8"
ipdb = "^0.12.2"
boto3 = "^1.9"
boto = "^2.49"
consoler = "^0.1.2"
whitenoise = "^5.0.1"
wagtail-cache = "^1.0.0"
django_select2 = "^7.4.2"




[tool.poetry.dev-dependencies]
# Django Dev Panel has an incompatibility with Wagtail-Cache, this is "fixed" in my fork, so we're installing from there until my PR gets merged.
django-developer-panel = {git = "https://github.com/drcongo/django-developer-panel.git", rev="55a6721e3df846a97f829a19374103037ba1cb29"}
coverage = "^4.5"
csscompressor = "^0.9.5"
cssmin = "^0.2.0"
factory-boy = "^2.11"
ipdb = "^0.12.2"
jsmin = "^2.2"
mixer = "^6.1"
pdbpp = "^0.9.3"
pytest = "^4.0"
pytest-cov = "^2.6"
pytest-django = "^3.4"
pytest-factoryboy = "^2.0"
pytest-pythonpath = "^0.7.3"
pytest-sugar = "^0.9.2"
rcssmin = "^1.0"
yuicompressor = "^2.4"
pudb = "^2018.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
